+++ psipy/tracing/tests/test_tracing.py
@@ -1,3 +1,4 @@
+import astropy.constants as const
 import astropy.units as u
 import numpy as np
 
@@ -23,7 +24,7 @@ def test_tracer():
 
     tracer = FortranTracer()
 
-    r = 40
+    r = 40 * const.R_sun
     lat = 0 * u.deg
     lon = 0 * u.deg
 
@@ -47,7 +48,7 @@ def test_fline_io(mas_model, tmpdir):
     model = MASOutput(mas_path)
     tracer = FortranTracer()
 
-    r = 40
+    r = 40 * const.R_sun
     lat = 0 * u.deg
     lon = 0 * u.deg