+++ CHANGELOG.rst
@@ -3,6 +3,11 @@ Changelog
 
 Version 0.3.1
 -------------
+Breaking changes
+~~~~~~~~~~~~~~~~
+- The radial coordinate input to `psipy.tracing.FortranTracer.trace` is now required to be
+  a `~astropy.units.Quantity` (with length units) instead of a plain numpy array.
+
 Bug fixes
 ~~~~~~~~~
 - Fixed the in-situ sampling example.